requirements.txt
# SmugDups v5.0 - SmugMug Duplicate Photo Manager
# Python package requirements

# GUI Framework
PyQt6>=6.4.0

# Data Processing
pandas>=1.5.0

# Image Processing
Pillow>=9.0.0

# HTTP Requests and OAuth
requests>=2.28.0
requests-oauthlib>=1.3.1

# Optional: Enhanced JSON handling
ujson>=5.4.0

# Optional: Progress bars for console
tqdm>=4.64.0

# Optional: Enhanced date parsing
python-dateutil>=2.8.2
